HCL/ai-research-agent/requirements.txt
streamlit
requests
pandas
plotly
reportlab
python-dotenv
